% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.libraries.R
\name{ensure.libraries}
\alias{ensure.libraries}
\title{Ensure that package libraries are installed and available.}
\usage{
ensure.libraries(packages.to.install = "tidyverse", verbose = FALSE)
}
\arguments{
\item{packages.to.install}{- a vector containing names of packages to install}

\item{verbose}{- a TRUE/FALSE value for verbose operation}
}
\value{
messages from library installs
}
\description{
Loads and attaches add-on packages. If the package is not present locally 
the package is installed from the CRAN repo. This is useful in making code run 
after new R installation or upon transfer to other machines.
}
\examples{
## simple installation of add-on packages
##
## require devtools to install package. 

 
library(devtools)
install_github("ww44ss/ensure.libraries")
library(ensure.libraries)

packages <- c("tidyverse",
              "ggplot2",
              "stringr",
              "ww44ss/yo"
              )

ensure.libraries(packages)

## can also use shorter form
ensure(packages)

## end (not run)
}
\keyword{packages}

