% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.libraries.R
\name{ensure}
\alias{ensure}
\title{ensure installs librarys and also checks for installed packages
and installs them if not already locally available.}
\usage{
ensure(packages.to.install = "tidyverse", verbose = FALSE)
}
\arguments{
\item{packages.to.install}{- a vector containing names of packages to install}

\item{verbose}{- a TRUE/FALSE value for verbose operation}
}
\value{
messages from library installs
}
\description{
ensure installs librarys and also checks for installed packages
and installs them if not already locally available.
}
\examples{
## simple installation of packages
##
## require devtools to install package. 
## then create a vector containing the names of needed packages. 
 
library(devtools)
github_install("ww44ss/ensure.libraries")

packages.needed <- c( 
                             "tidyverse",
                             "gglot2",
                             "stringr",
                             "ww44ss/yo"
                           )

ensure.libraries(packages.needed)

## can also use shorter form
ensure(libraries.needed)

}
\keyword{packages}

